requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Image processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Machine Learning
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
diffusers==0.21.4
accelerate==0.24.1
safetensors==0.4.0

# MediaPipe for pose/face detection
mediapipe==0.10.7

# Redis for task queue
redis==5.0.1
aioredis==2.0.1

# Storage
boto3==1.34.0
botocore==1.34.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0